hop-node: DRY route config filters for db getter
